refactor: RoomController와 RoomApiController 분리
