+++ src/main/resources/static/login.html
@@ -15,14 +15,14 @@
         <img src="images/black_k.png" class="icon-image">
     </div>
     <form action="http://localhost:8080/rooms" onsubmit="return checkNameNotEmpty();" method="post">
-        <input type="text" id="name" name="name" placeholder="방 이름" class="input-name">
+        <input type="text" id="name" name="name" placeholder="방 이름" class="input-name" required>
         <input type="password" id="password" name="password" placeholder="비밀번호" class="input-name" required>
         <br>
         <button type="submit" class="input-button">생성</button>
     </form>
     <span id="error-message"></span>
 
-    <button onclick="fetchRoomNames()">방 목록</button>
+    <br><button onclick="fetchRoomNames()">방 목록</button>
     <div id="roomNames" class="room-names"></div>
 </div>
 </body>
@@ -37,7 +37,7 @@
     }
 
     function fetchRoomNames() {
-        fetch("http://localhost:8080/rooms", {
+        fetch("http://localhost:8080/api/rooms", {
             method: 'GET',
             headers: {
                 'Content-type': 'application/json; charset=UTF-8',
@@ -72,7 +72,7 @@
             });
 
         function fetchDeleteRoom(roomId, password) {
-            fetch("http://localhost:8080/rooms/" + roomId, {
+            fetch("http://localhost:8080/api/rooms/" + roomId, {
                 method: 'DELETE',
                 headers: {
                     'Content-type': 'application/x-www-form-urlencoded; charset=UTF-8',
@@ -80,8 +80,14 @@
                 body: 'password=' + password
             })
                 .then(res =>{
-                    let roomContainer = document.getElementById('roomNames');
-                    roomContainer.innerHTML = "";
+                    let errorMessage = document.getElementById('error-message');
+                    errorMessage.innerHTML = "";
+
+                    if (res.status === 400) {
+                        errorMessage.innerHTML = '비밀번호가 일치하지 않습니다.';
+                    }
+
+                    fetchRoomNames();
                 })
         }
     }
